anamnese-system/requirements.txt
PyMuPDF
opencv-python
Pillow
pytesseract
fastapi
uvicorn[standard]
python-multipart
pytest
